+++ prettier.config.js
@@ -1,3 +1,3 @@
-import config from '@solana/ms-tools-config/prettier.config.js'
+import config from '@solana-foundation/ms-tools-config/prettier.config.js'
 
 export default config